+++ piecash2/__init__.py
@@ -1,3 +1,5 @@
-from .schema.schema_generation import open_book, add_book_module_in_path
+from piecash2.schema.generation.schema_generation import add_book_module_in_path
+
+from .core.book import open_book
 
 __ALL__ = [open_book, add_book_module_in_path]
