+++ piecash2/schema/generation/schema_generation.py
@@ -3,10 +3,9 @@
 import importlib.util
 import sqlite3
 import sys
+import textwrap
+from datetime import datetime
 from pathlib import Path
-
-import sqlalchemy
-import sqlalchemy.orm
 
 # folder with code to insert into the schema generated by sqlacodegen
 code_templates = Path(__file__).parent
@@ -17,6 +16,7 @@ path_schemas.mkdir(exist_ok=True, parents=True)
 
 
 def import_module_from_path(path: Path):
+    """Import a module from a path."""
     module_name = path.stem
     spec = importlib.util.spec_from_file_location(module_name, path)
     module = importlib.util.module_from_spec(spec)  # type: ignore[arg-type]
@@ -26,6 +26,7 @@ def import_module_from_path(path: Path):
 
 
 def import_gnucash(book, regenerate_schema=False):
+    """Import the gnucash schema from the sqlite database as a module. If the schema does not exist, generate it."""
     modulename = get_schema_name(book)
 
     if not modulename.exists() or regenerate_schema:
@@ -51,8 +52,20 @@ def generate_schema(db, schema_file: Path):
 
     # insert before/after the generated schema the common schema
     schema_file_text = (
-        (code_templates / "sa_schema_pre.py").read_text()
+        textwrap.dedent(
+            """
+        # -*- coding: utf-8 -*-
+        '''This file has been generated by piecash2.schema.generation.generate_schema on {datetime.now().isoformat()}.
+        '''
+        """
+        )
+        + (code_templates / "sa_schema_pre.py").read_text()
         + schema_file.read_text()
+        .replace(
+            "from sqlalchemy.ext.declarative import declarative_base as declarative_base_",
+            "from sqlalchemy.orm import declarative_base as declarative_base_",
+        )
+        .replace("from sqlalchemy.orm import Mapped, declarative_base, mapped_column", "from sqlalchemy.orm import Mapped, mapped_column")
         + (code_templates / "sa_schema_post.py").read_text()
         + ""
     )
@@ -67,6 +80,7 @@ def get_version(db: Path):
         c = conn.cursor()
         c.execute("SELECT table_version FROM versions WHERE table_name=='Gnucash'")
         (gnucash_version,) = c.fetchone()
+
     return gnucash_version
 
 
@@ -76,36 +90,15 @@ def get_schema_name(book: Path):
     return path_schemas / f"book_schema_sqlite_{v}.py"
 
 
-def open_book(book, regenerate_schema=False):
-    if isinstance(book, str):
-        book = Path(book)
-
-    book_posix = book.as_posix()
-
-    # make a backup of the DB in memory
-    db_memory_name = f":memgeco_{abs(hash(book_posix))}:"
-    with sqlite3.connect(book_posix) as source:
-        sqliteconn = f"file:{db_memory_name}?mode=memory&cache=shared"
-        dest = sqlite3.connect(sqliteconn, uri=True)
-        source.backup(dest)
-
-    engine = sqlalchemy.create_engine(f"sqlite:///{db_memory_name}", echo=False, creator=lambda: sqlite3.connect(sqliteconn, uri=True))
-
-    Session = sqlalchemy.orm.sessionmaker(bind=engine, autoflush=True, autocommit=False)
-
-    # must execute some query otherwise future call to the Session raise error
-    with Session() as s:
-        s.execute(sqlalchemy.text(""))
-
-    Session.module = import_gnucash(book, regenerate_schema=regenerate_schema)
-
-    return Session
-
-
 def add_book_module_in_path(book):
     module_folder = str(get_schema_name(book).parent)
     if module_folder not in sys.path:
         sys.path.append(module_folder)
+
+
+def unload_module(mod):
+    if mod in sys.modules:
+        del sys.modules[mod]
 
 
 def remove_book_module_in_path(book):
@@ -113,7 +106,7 @@ def remove_book_module_in_path(book):
     module_folder = str(module_path.parent)
     module_name = module_path.stem
 
-    sys.modules.pop(module_name)
+    unload_module(module_name)
 
     if module_folder in sys.path:
         sys.path.remove(module_folder)
